requirements.txt
Flask~=3.0.2
Flask-WTF==0.15.1
seaborn~=0.13.2
matplotlib~=3.7.5
opencv-python~=4.9.0.80
Pillow~=10.2.0
Werkzeug~=3.0.1
wtforms~=3.1.2
Flask-Recaptcha==0.5.1
numpy~=1.24.4
